Import ready: create util method for converting platform name
